truetandem/e-QIP-prototype#1476 Moved our foreign passport into its own subsection
