+++ src/components/ProgressBar/ProgressBar.test.jsx
@@ -29,7 +29,8 @@ describe('The progress bar component', () => {
       Completed: {
         citizenship: [
           { section: 'citizenship', subsection: 'status', valid: true },
-          { section: 'citizenship', subsection: 'multiple', valid: true }
+          { section: 'citizenship', subsection: 'multiple', valid: true },
+          { section: 'citizenship', subsection: 'passorts', valid: true }
         ]
       }
     }
