+++ src/validators/citizenship-multiple.test.js
@@ -1,277 +1,21 @@
-import CitizenshipMultipleValidator, { CitizenshipItemValidator, PassportItemValidator, TravelValidator } from './citizenship-multiple'
+import CitizenshipMultipleValidator, { CitizenshipItemValidator } from './citizenship-multiple'
 import { battery } from './helpers'
 import Location from '../components/Form/Location'
 
 describe('citizenship multiple component validation', function () {
-  it('can validate travel country', () => {
-    const tests = [
-      {
-        state: {
-          Country: ''
-        },
-        expected: false
-      },
-      {
-        state: {
-          Country: 'United States'
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, TravelValidator, 'validCountry')
-  })
-
-  it('can validate travel dates', () => {
-    const tests = [
-      {
-        state: {
-          Dates: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Dates: {
-            from: {
-              date: new Date('1/1/2010')
-            },
-            to: {
-              date: new Date('1/1/2012')
-            },
-            present: false
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, TravelValidator, 'validDates')
-  })
-
-  it('can validate passport country', () => {
-    const tests = [
-      {
-        state: {
-          Country: ''
-        },
-        expected: false
-      },
-      {
-        state: {
-          Country: 'United States'
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validCountry')
-  })
-
-  it('can validate passport date issued', () => {
-    const tests = [
-      {
-        state: {
-          Issued: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Issued: {
-            day: '1',
-            month: '1',
-            year: '2016',
-            date: new Date('1/1/2016')
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validIssued')
-  })
-
-  it('can validate passport location issued', () => {
-    const tests = [
-      {
-        state: {
-          Location: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Location: {
-            country: 'United States',
-            city: 'Arlington',
-            layout: Location.CITY_COUNTRY
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validLocation')
-  })
-
-  it('can validate passport name', () => {
-    const tests = [
-      {
-        state: {
-          Name: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Name: {
-            first: 'Foo',
-            firstInitialOnly: false,
-            middle: 'J',
-            middleInitialOnly: true,
-            noMiddleName: false,
-            last: 'Bar',
-            lastInitialOnly: false,
-            suffix: 'Jr'
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validName')
-  })
-
-  it('can validate passport number', () => {
-    const tests = [
-      {
-        state: {
-          Number: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Number: {
-            value: 'number'
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validNumber')
-  })
-
-  it('can validate passport date of expiration', () => {
-    const tests = [
-      {
-        state: {
-          Expiration: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Expiration: {
-            day: '1',
-            month: '1',
-            year: '2016',
-            date: new Date('1/1/2016')
-          }
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validExpiration')
-  })
-
-  it('can validate passport usage', () => {
-    const tests = [
-      {
-        state: {
-          Used: null
-        },
-        expected: false
-      },
-      {
-        state: {
-          Used: 'Yuppers'
-        },
-        expected: false
-      },
-      {
-        state: {
-          Used: 'No'
-        },
-        expected: true
-      },
-      {
-        state: {
-          Used: 'Yes'
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validUsed')
-  })
-
-  it('can validate passport travel usage', () => {
-    const tests = [
-      {
-        state: {
-          Used: 'No'
-        },
-        expected: true
-      },
-      {
-        state: {
-          Used: 'Yes',
-          Countries: []
-        },
-        expected: false
-      },
-      {
-        state: {
-          Used: 'Yes',
-          Countries: [
-            {
-              Item: {
-                Country: 'United States',
-                Dates: {
-                  from: {
-                    date: new Date('1/1/2010')
-                  },
-                  to: {
-                    date: new Date('1/1/2012')
-                  },
-                  present: false
-                }
-              }
-            }
-          ]
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, PassportItemValidator, 'validCountries')
-  })
-
   it('can validate citizenship country', () => {
     const tests = [
       {
         state: {
-          Country: ''
+          Country: {}
         },
         expected: false
       },
       {
         state: {
-          Country: 'United States'
+          Country: {
+            value: 'United States'
+          }
         },
         expected: true
       }
@@ -441,7 +185,9 @@ describe('citizenship multiple component validation', function () {
           Citizenships: [
             {
               Item: {
-                Country: 'United States',
+                Country: {
+                  value: 'United States'
+                },
                 Dates: {
                   from: {
                     date: new Date('1/1/2010')
@@ -474,153 +220,6 @@ describe('citizenship multiple component validation', function () {
     battery(tests, CitizenshipMultipleValidator, 'validCitizenships')
   })
 
-  it('can validate multiple passports', () => {
-    const tests = [
-      {
-        state: {
-          Passports: [
-            {
-              Item: {}
-            }
-          ]
-        },
-        expected: false
-      },
-      {
-        state: {
-          Passports: [
-            {
-              Has: 'No'
-            }
-          ]
-        },
-        expected: true
-      },
-      {
-        state: {
-          Passports: [
-            {
-              Has: 'Yes',
-              Item: {
-                Country: 'United States',
-                Issued: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Location: {
-                  country: 'United States',
-                  city: 'Arlington',
-                  layout: Location.CITY_COUNTRY
-                },
-                Name: {
-                  first: 'Foo',
-                  firstInitialOnly: false,
-                  middle: 'J',
-                  middleInitialOnly: true,
-                  noMiddleName: false,
-                  last: 'Bar',
-                  lastInitialOnly: false,
-                  suffix: 'Jr'
-                },
-                Number: {
-                  value: 'number'
-                },
-                Expiration: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Used: 'Yes',
-                Countries: [
-                  {
-                    Item: {
-                      Country: '',
-                      Dates: {
-                        from: {
-                          date: new Date('1/1/2010')
-                        },
-                        to: {
-                          date: new Date('1/1/2012')
-                        },
-                        present: false
-                      }
-                    }
-                  }
-                ]
-              }
-            }
-          ]
-        },
-        expected: false
-      },
-      {
-        state: {
-          Passports: [
-            {
-              Has: 'Yes',
-              Item: {
-                Country: 'United States',
-                Issued: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Location: {
-                  country: 'United States',
-                  city: 'Arlington',
-                  layout: Location.CITY_COUNTRY
-                },
-                Name: {
-                  first: 'Foo',
-                  firstInitialOnly: false,
-                  middle: 'J',
-                  middleInitialOnly: true,
-                  noMiddleName: false,
-                  last: 'Bar',
-                  lastInitialOnly: false,
-                  suffix: 'Jr'
-                },
-                Number: {
-                  value: 'number'
-                },
-                Expiration: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Used: 'Yes',
-                Countries: [
-                  {
-                    Item: {
-                      Country: 'United States',
-                      Dates: {
-                        from: {
-                          date: new Date('1/1/2010')
-                        },
-                        to: {
-                          date: new Date('1/1/2012')
-                        },
-                        present: false
-                      }
-                    }
-                  }
-                ]
-              }
-            }
-          ]
-        },
-        expected: true
-      }
-    ]
-
-    battery(tests, CitizenshipMultipleValidator, 'validPassports')
-  })
-
   it('can validate entire multiple citizenships and passports', () => {
     const tests = [
       {
@@ -629,7 +228,9 @@ describe('citizenship multiple component validation', function () {
           Citizenships: [
             {
               Item: {
-                Country: 'United States',
+                Country: {
+                  value: 'United States'
+                },
                 Dates: {
                   from: {
                     date: new Date('1/1/2010')
@@ -653,62 +254,7 @@ describe('citizenship multiple component validation', function () {
               }
             }
           ],
-          CitizenshipsBranch: 'No',
-          Passports: [
-            {
-              Has: 'Yes',
-              Item: {
-                Country: 'United States',
-                Issued: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Location: {
-                  country: 'United States',
-                  city: 'Arlington',
-                  layout: Location.CITY_COUNTRY
-                },
-                Name: {
-                  first: 'Foo',
-                  firstInitialOnly: false,
-                  middle: 'J',
-                  middleInitialOnly: true,
-                  noMiddleName: false,
-                  last: 'Bar',
-                  lastInitialOnly: false,
-                  suffix: 'Jr'
-                },
-                Number: {
-                  value: 'number'
-                },
-                Expiration: {
-                  day: '1',
-                  month: '1',
-                  year: '2016',
-                  date: new Date('1/1/2016')
-                },
-                Used: 'Yes',
-                Countries: [
-                  {
-                    Item: {
-                      Country: 'United States',
-                      Dates: {
-                        from: {
-                          date: new Date('1/1/2010')
-                        },
-                        to: {
-                          date: new Date('1/1/2012')
-                        },
-                        present: false
-                      }
-                    }
-                  }
-                ]
-              }
-            }
-          ]
+          CitizenshipsBranch: 'No'
         },
         expected: true
       }
