+++ src/validators/index.js
@@ -35,6 +35,7 @@ import PsychologicalValidator from './psychological'
 import CompetenceValidator from './competence'
 import CitizenshipValidator from './citizenship'
 import CitizenshipMultipleValidator from './citizenship-multiple'
+import CitizenshipPassportsValidator from './citizenship-passports'
 import CivilUnionValidator from './civilunion'
 import MaritalValidator from './marital'
 import CohabitantsValidator from './cohabitant'
@@ -125,6 +126,7 @@ export {
   HospitalizationValidator,
   CitizenshipValidator,
   CitizenshipMultipleValidator,
+  CitizenshipPassportsValidator,
   CivilUnionValidator,
   MaritalValidator,
   CohabitantsValidator,
